Modified to work better with fast devices like ESP32
The TM1638 is basically a slow SPI device (< 500kHz) in DIO mode.
The clock uses the equivalent of SPI mode 3 (normally high, clocks data on
the rising edge). The shift clocking is slowed with a small delay (on the
order of a microsecond) and inverted accordingly.

Note: The delay has more of an impact than the clock sense appears to for
the devices I've tested.